batch tagging (and mass tag updating to a generic function)
